Hackathon/frontend/requirements.txt
streamlit==1.27.2
plotly==5.17.0
requests==2.31.0
pandas==2.1.1
Pillow==10.0.1
